hack: Switch from using session to using cookies so the data will fit
Rails' cookie_store session object can only hold so much data. Because
we're putting massive certs / keys in session it runs out of room. To
work around this in a very hacky way I've stored every bit of user input
in a separate cookie (:troll:).

This would be much better if we used a database as a session store, but
I can't be bothered to set that up.

Authors: @richardtowers